Treat improper uses of "" in filtegraphs as errors
Filtergraohs are expanded as if hey are always enclosed in a pair of double quotes, so that exact spacing is preserved. However, if the user places additional, balanced double quotes, the expansion will result in more than one "word" and the spacing between he two will be lost.

A check is added so that if word expansion produces more than one word, an error is singalled.